Added init section and in runtime time inputs
The global configurations are defined in a block with a "# INIT" comment
in the beginning. It's recommendable to place this block as the first,
because only the next blocks will be affected for the configurations
written in it.
It's possible to specify inputs names (keys whose value will be asked to
the user in execution time), being able to use them in anywhere in the
rest of the script.